(packaging) temporary workaround to get PE fork building
We just created this fork, and many, many things will need to get fixed in
packaging to undo its foss/pe hybrid model. But for now, just changing the PE
conditional and adding build_pe as true to the build_defaults will suffice to
get it working with our packaging pipeline.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>