(maint) Update logrotate munging to use realsysconfdir
Previously puppetdb's logrotate munging munged a file rooted in sysconfdir.
This works in foss, where sysconfdir is etc, but breaks in PE, where sysconfdir
is /etc/puppetlabs. This commit updates the spec to use realsysconfdir instead
of sysconfdir, to ensure we try to munge /etc/logrotate.d/pe-puppetdb and not
/etc/puppetlabs/logrotate.d/pe-puppetdb (which won't exist).