(maint) Add dev level pretty SQL tracing
This patch adds a dev only feature to be used primarily via an nrepl to
enable both pretty SQL and explain outputs to be printed to the info log
output. Useful for debugging live API queries for performance hangups, and
modifying code on the fly to see how it affects performance.

Signed-off-by: Ken Barber <ken@bob.sh>